+++ src/include/LSP/Workspace.hpp
@@ -9,6 +9,14 @@
 #include "LSP/Client.hpp"
 #include "LSP/WorkspaceFileResolver.hpp"
 
+struct Reference
+{
+    const Luau::ModuleName moduleName;
+    const Luau::Location location;
+    // NOTE: DO NOT STORE THIS EXPR, it will be freed if the module becomes dirty
+    const Luau::AstExpr* expr = nullptr;
+};
+
 class WorkspaceFolder
 {
 public:
@@ -63,6 +71,8 @@ private:
         const TextDocument& textDocument, std::vector<lsp::CompletionItem>& result);
     lsp::WorkspaceEdit computeOrganiseRequiresEdit(const lsp::DocumentUri& uri);
     lsp::WorkspaceEdit computeOrganiseServicesEdit(const lsp::DocumentUri& uri);
+    std::optional<std::vector<Reference>> WorkspaceFolder::findAllReferences(
+        const Luau::TypeId ty, std::optional<Luau::Name> property = std::nullopt);
 
 public:
     std::vector<std::string> getComments(const Luau::ModuleName& moduleName, const Luau::Location& node);